First Commit
Commit initial design files. This design printed and flew, but has some
issues. The arms are too wobbly in the Z direction. The frame is too
heavy. The battery attachment is awkward. The camera mount could allow
some more room for the camera. The motor holders could be a bit longer.